pyproject.toml
[tool.poetry]
name = "goit-pyweb-hw-11"
version = "0.1.0"
description = ""
authors = ["MaxRevan <gladkovnissan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.4"}
pydantic-settings = "^2.6.1"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
asyncpg = "^0.30.0"
passlib = "^1.7.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
fastapi-mail = "^1.4.1"
fastapi-limiter = "^0.1.6"
fastapi-cache2 = "^0.2.2"
bcrypt = "<4.0.0"
libgravatar = "^1.0.4"
cloudinary = "^1.41.0"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pytest-faker = "^2.0.0"
pytest-cov = "^6.0.0"
greenlet = "^3.1.1"


[tool.poetry.group.dev.dependencies]
sphinx = "^8.1.3"

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
